Update Greek translation and fix issues
The following changes were made:

* Updated the Greek translation. Note that as I was unable to test certain functionality in my Xera installation, I could not verify some strings. I might have to make an account with a Xera-powered host to test out and update if required. Updated English with small changes as well.

* Updated URL issues caused by the move of the assets to a /default folder, which had not been properly updated.

* Restored the changes in #29, which had been overwritten by the move to a language system.

I am uncertain about the change of "some reason" to {some_reason}, while the former did not fit well with me, I'm not sure if the latter will display the actual reason. If you can confirm, or update it if it does not, I'll be grateful.

Huge thanks once again to @bopv2 for the translation.